polls/requirements.txt
asgiref==3.7.2
Django==4.2.11
django-rest-framework==0.1.0
mysql-client==0.0.1
PyMySQL==1.1.0
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.10.0
